+++ src/main/java/com/hhplu/hhplusconcert/app/application/service/concert/service/SeatService.java
@@ -50,4 +50,10 @@ public class SeatService {
             seat.changeStatus(SeatStatus.RESERVED);
         }
     }
+
+    public long calculateTotalPoint(List<Seat> seats) {
+        return seats.stream()
+                .mapToLong(Seat::getSeatPrice)
+                .sum();
+    }
 }
