+++ src/test/java/com/hhplu/hhplusconcert/app/application/scheduler/WaitingQueueStatusSchedulerTest.java
@@ -1,6 +1,6 @@
 package com.hhplu.hhplusconcert.app.application.scheduler;
 
-import com.hhplu.hhplusconcert.app.application.service.waitingqueue.service.WaitingQueueRedisService;
+import com.hhplu.hhplusconcert.app.application.service.waitingqueue.WaitingQueueRedisService;
 import com.hhplu.hhplusconcert.app.domain.user.entity.User;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
