+++ src/test/java/com/hhplu/hhplusconcert/app/unittest/facade/ReservationFacadeTest.java
@@ -3,10 +3,10 @@ package com.hhplu.hhplusconcert.app.unittest.facade;
 import com.hhplu.hhplusconcert.app.application.facade.ReservationFacade;
 import com.hhplu.hhplusconcert.app.application.service.concert.service.ConcertService;
 import com.hhplu.hhplusconcert.app.application.service.concert.service.SeatService;
-import com.hhplu.hhplusconcert.app.application.service.payment.service.PaymentService;
+import com.hhplu.hhplusconcert.app.application.service.payment.PaymentService;
 import com.hhplu.hhplusconcert.app.application.service.reservation.dto.ReserveSeatsDto;
 import com.hhplu.hhplusconcert.app.application.service.reservation.dto.ReserveSeatsResponseDto;
-import com.hhplu.hhplusconcert.app.application.service.reservation.service.ReservationService;
+import com.hhplu.hhplusconcert.app.application.service.reservation.ReservationService;
 import com.hhplu.hhplusconcert.app.domain.concert.entity.Concert;
 import com.hhplu.hhplusconcert.app.domain.concert.entity.Seat;
 import com.hhplu.hhplusconcert.app.domain.reservation.entity.Reservation;
