+++ src/test/java/com/hhplu/hhplusconcert/app/unittest/service/ReservationServiceUnitTest.java
@@ -1,6 +1,6 @@
 package com.hhplu.hhplusconcert.app.unittest.service;
 
-import com.hhplu.hhplusconcert.app.application.service.reservation.service.ReservationService;
+import com.hhplu.hhplusconcert.app.application.service.reservation.ReservationService;
 import com.hhplu.hhplusconcert.app.domain.reservation.ReservationStatus;
 import com.hhplu.hhplusconcert.app.domain.reservation.entity.Reservation;
 import com.hhplu.hhplusconcert.app.domain.reservation.repository.ReservationRepository;
